pred realtime
