+++ src/autogluon/cloud/backend/constant.py
@@ -1,3 +1,5 @@
 SAGEMAKER = "sagemaker"
+TABULAR_SAGEMAKER = "tabular_sagemaker"
+MULTIMODL_SAGEMAKER = "multimodal_sagemaker"
 TIMESERIES_SAGEMAKER = "timeseries_sagemaker"
 RAY = "ray"
